First pass transcription c56-0081
Relates to issue #112. No special encoding of chapter heading—I don’t
know of any decision on how to encode these. “Illuminating” appeared
overwritten to me, so I did my best with that encoding. I suspect there
needs to be addition of <mod> to this file but I’m not clear on how to
determine the extent of these modifications.